wire in the CONSUME framework to start with the subscriptions
