+++ src/main/java/io/mapsmessaging/network/protocol/impl/nats/jetstream/stream/consumer/ConsumerManager.java
@@ -1,22 +1,25 @@
 package io.mapsmessaging.network.protocol.impl.nats.jetstream.stream.consumer;
 
-import io.mapsmessaging.network.protocol.impl.nats.frames.NatsFrame;
-import io.mapsmessaging.network.protocol.impl.nats.frames.PayloadFrame;
+import io.mapsmessaging.network.protocol.impl.nats.jetstream.stream.JetStreamFrameHandler;
 import io.mapsmessaging.network.protocol.impl.nats.jetstream.stream.RequestHandler;
-import io.mapsmessaging.network.protocol.impl.nats.state.SessionState;
-
-import java.io.IOException;
+import io.mapsmessaging.network.protocol.impl.nats.jetstream.stream.consumer.handler.*;
 
 public class ConsumerManager extends RequestHandler {
+
+  public ConsumerManager() {
+    super(new JetStreamFrameHandler[]{
+        new CreateHandler(),
+        new DeleteHandler(),
+        new InfoHandler(),
+        new ListHandler(),
+        new MessageHandler(),
+        new NamesHandler()
+    });
+  }
 
   @Override
   public String getType() {
     return "$JS.API.CONSUMER";
-  }
-
-  @Override
-  public NatsFrame process(String subject, PayloadFrame frame, SessionState sessionState) throws IOException {
-    return null;
   }
 
 }
