+++ src/main/java/io/mapsmessaging/network/protocol/impl/nats/jetstream/stream/consumer/handler/MessageHandler.java
@@ -1,5 +1,4 @@
-package io.mapsmessaging.network.protocol.impl.nats.jetstream.stream.api.handlers;
-
+package io.mapsmessaging.network.protocol.impl.nats.jetstream.stream.consumer.handler;
 
 import com.google.gson.JsonObject;
 import io.mapsmessaging.network.protocol.impl.nats.frames.NatsFrame;
@@ -9,14 +8,16 @@ import io.mapsmessaging.network.protocol.impl.nats.state.SessionState;
 
 import java.io.IOException;
 
-public class StreamPurgeHandler extends JetStreamFrameHandler {
+public class MessageHandler  extends JetStreamFrameHandler {
+
   @Override
   public String getName() {
-    return "PURGE";
+    return "MSG.NEXT";
   }
 
   @Override
   public NatsFrame handle(PayloadFrame frame, JsonObject json, SessionState sessionState) throws IOException {
     return null;
   }
+
 }
